+++ buildSrc/src/main/kotlin/Config.kt
@@ -17,7 +17,7 @@
 
 object Realm {
     val ciBuild = (System.getenv("JENKINS_HOME") != null)
-    const val version = "1.1.1-SNAPSHOT"
+    const val version = "1.2.0"
     const val group = "io.realm.kotlin"
     const val projectUrl = "https://realm.io"
     const val pluginPortalId = "io.realm.kotlin"
@@ -53,11 +53,11 @@ object Realm {
 object Versions {
     object Android {
         const val minSdk = 16
-        const val targetSdk = 31
-        const val compileSdkVersion = 31
-        const val buildToolsVersion = "31.0.0"
-        const val buildTools = "7.1.0" // https://maven.google.com/web/index.html?q=gradle#com.android.tools.build:gradle
-        const val ndkVersion = "23.1.7779620"
+        const val targetSdk = 33
+        const val compileSdkVersion = 33
+        const val buildToolsVersion = "33.0.0"
+        const val buildTools = "7.2.2" // https://maven.google.com/web/index.html?q=gradle#com.android.tools.build:gradle
+        const val ndkVersion = "23.2.8568313"
     }
     const val androidxBenchmarkPlugin = "1.1.0-beta04" // https://maven.google.com/web/index.html#androidx.benchmark:androidx.benchmark.gradle.plugin
     const val androidxStartup = "1.1.0" // https://maven.google.com/web/index.html?q=startup#androidx.startup:startup-runtime
@@ -68,7 +68,7 @@ object Versions {
     const val autoService = "1.0" // https://mvnrepository.com/artifact/com.google.auto.service/auto-service
     // Not currently used, so mostly here for documentation. Core requires minimum 3.15, but 3.18.1 is available through the Android SDK.
     // Build also tested successfully with 3.21.4 (latest release).
-    const val cmake = "3.18.1"
+    const val cmake = "3.22.1"
     const val coroutines = "1.6.0-native-mt" // https://mvnrepository.com/artifact/org.jetbrains.kotlinx/kotlinx-coroutines-core
     const val datetime = "0.3.2" // https://github.com/Kotlin/kotlinx-datetime
     const val detektPlugin = "1.19.0-RC1" // https://github.com/detekt/detekt
@@ -80,8 +80,8 @@ object Versions {
     const val jvmTarget = "1.8"
     // When updating the Kotlin version, also remember to update /examples/min-android-sample/build.gradle.kts
     const val kotlin = "1.6.10" // https://github.com/JetBrains/kotlin and https://kotlinlang.org/docs/releases.html#release-details
-    const val latestKotlin = "1.7.0-RC" // https://kotlinlang.org/docs/eap.html#build-details
-    const val kotlinCompileTesting = "1.4.2" // https://github.com/tschuchortdev/kotlin-compile-testing
+    const val latestKotlin = "1.7.20-RC" // https://kotlinlang.org/docs/eap.html#build-details
+    const val kotlinCompileTesting = "1.4.9" // https://github.com/tschuchortdev/kotlin-compile-testing
     const val ktlint = "0.45.2" // https://github.com/pinterest/ktlint
     const val ktor = "1.6.8" // https://github.com/ktorio/ktor
     const val nexusPublishPlugin = "1.1.0" // https://github.com/gradle-nexus/publish-plugin